Changes to Rick Exhibit
Made a prominent hover effect on Rick's Name at the top of each gallery.
This is the primary link back to the Rick Home page.

Added a hamburger menu for small screens that gives access to site
navigation.

Added direct links to the about and contact pages from the footer.

Added link to current exhibit in the contact and about page hamburger
menu