Changes to Rick gallery, home and about pages
Changed the icons for the video thumbnails. added a final slide to each
gallery with a link to the Rick Home page and to some of rick's cool
music links.
changed the welcome panel on the home page to a lighter grey to give me
more flexibility with text and link colors.
changed the text on the about page