add 3 new pots to jug a joe gallery
I also tries to minify my html. I added the vecessary script in my
gulpfile and installed the needed npm scripts. It made my php code fail
because a php comment is just everything after a //. So after being
minfied everything after the first php comment was commented out. So I
removed this code from my gulpfile. but that is why their are so many
node_modules in this commit.